LibWeb: Split AbstractImageStyleValue out of StyleValue.{h,cpp}
