+++ Userland/Libraries/LibWeb/CSS/StyleValues/ImageStyleValue.h
@@ -11,7 +11,7 @@
 
 #include <AK/URL.h>
 #include <LibWeb/CSS/Enums.h>
-#include <LibWeb/CSS/StyleValue.h>
+#include <LibWeb/CSS/StyleValues/AbstractImageStyleValue.h>
 
 namespace Web::CSS {
