+++ Userland/Libraries/LibWeb/CSS/StyleValues/RadialGradientStyleValue.h
@@ -11,7 +11,7 @@
 
 #include <AK/Vector.h>
 #include <LibWeb/CSS/Enums.h>
-#include <LibWeb/CSS/StyleValue.h>
+#include <LibWeb/CSS/StyleValues/AbstractImageStyleValue.h>
 
 namespace Web::CSS {
